broke tracker out from GUI
